@@ -5,8 +5,8 @@ version: 0.0.1
 publish_to: none
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: 3.3.10
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.10.1"
 
 dependencies:
   flutter:
@@ -19,14 +19,13 @@ dependencies:
     git:
       url: git@github.com:regentmarkets/flutter-deriv-packages.git
       path: packages/deriv_technical_analysis
-      ref: dev
+      ref: flutter-version-3
 
-  collection: ^1.15.0-nullsafety.4
   equatable: ^2.0.5
-  intl: ^0.17.0
-  json_annotation: ^4.7.0
+  intl: ^0.18.0
+  json_annotation: ^4.8.0
   provider: ^6.0.5
-  shared_preferences: ^2.0.15
+  shared_preferences: ^2.1.0
 
 dev_dependencies:
   flutter_test:
@@ -36,11 +35,11 @@ dev_dependencies:
     git:
       url: git@github.com:regentmarkets/flutter-deriv-packages.git
       path: packages/deriv_lint
-      ref: dev
+      ref: flutter-version-3
 
   build_runner: ^2.3.3
-  intl_utils: ^2.0.0
-  json_serializable: ^6.5.4
+  intl_utils: ^2.8.2
+  json_serializable: ^6.6.1
 
 flutter:
   assets:
