@@ -52,7 +52,7 @@ class ContinuousDrawingToolItemState
   Widget _buildColorField() => Row(
         children: <Widget>[
           Text(
-            ChartLocalization.of(context)!.labelColor,
+            ChartLocalization.of(context).labelColor,
             style: const TextStyle(fontSize: 16),
           ),
           ColorSelector(
