@@ -54,7 +54,7 @@ class VerticalDrawingToolItemState
   Widget _buildColorField() => Row(
         children: <Widget>[
           Text(
-            ChartLocalization.of(context)!.labelColor,
+            ChartLocalization.of(context).labelColor,
             style: const TextStyle(fontSize: 16),
           ),
           ColorSelector(
