@@ -1,21 +1,16 @@
-import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/creator.dart';
+import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/continuous/continuous_line_drawing.dart';
+import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/drawing_creator.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/data_model/edge_point.dart';
-import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/line/line_drawing.dart';
 import 'package:flutter/material.dart';
 import '../data_model/drawing_parts.dart';
 
 /// Creates a Continuous drawing piece by piece collected on every gesture
 /// exists in a widget tree starting from selecting a continuous drawing tool
 /// and until drawing should be finished.
-class ContinuousDrawingCreator extends Creator<LineDrawing> {
+class ContinuousDrawingCreator extends DrawingCreator<ContinuousLineDrawing> {
   /// Initializes the continuous drawing creator.
   const ContinuousDrawingCreator({
-    required void Function(
-      Map<String, List<LineDrawing>>, {
-      bool isDrawingFinished,
-      bool isInfiniteDrawing,
-    })
-        onAddDrawing,
+    required OnAddDrawing<ContinuousLineDrawing> onAddDrawing,
     required double Function(double) quoteFromCanvasY,
     required this.clearDrawingToolSelection,
     required this.removeDrawing,
@@ -38,12 +33,15 @@ class ContinuousDrawingCreator extends Creator<LineDrawing> {
   final bool shouldStopDrawing;
 
   @override
-  CreatorState<LineDrawing> createState() => _ContinuousDrawingCreatorState();
+  DrawingCreatorState<ContinuousLineDrawing> createState() =>
+      _ContinuousDrawingCreatorState();
 }
 
-class _ContinuousDrawingCreatorState extends CreatorState<LineDrawing> {
+class _ContinuousDrawingCreatorState
+    extends DrawingCreatorState<ContinuousLineDrawing> {
   @override
   void onTap(TapUpDetails details) {
+    super.onTap(details);
     final ContinuousDrawingCreator _widget = widget as ContinuousDrawingCreator;
 
     if (_widget.shouldStopDrawing) {
@@ -54,6 +52,8 @@ class _ContinuousDrawingCreatorState extends CreatorState<LineDrawing> {
     setState(() {
       position = details.localPosition;
       tapCount++;
+      final int currentTap = tapCount - 1;
+      final int previousTap = tapCount - 2;
 
       if (edgePoints.isEmpty) {
         /// Draw the initial point of the continuous.
@@ -61,9 +61,8 @@ class _ContinuousDrawingCreatorState extends CreatorState<LineDrawing> {
           epoch: epochFromX!(position!.dx),
           quote: widget.quoteFromCanvasY(position!.dy),
         ));
-        drawingId = 'continuous_${edgePoints.first.epoch}';
 
-        drawingParts.add(LineDrawing(
+        drawingParts.add(ContinuousLineDrawing(
           drawingPart: DrawingParts.marker,
           startEdgePoint: edgePoints.first,
         ));
@@ -71,8 +70,6 @@ class _ContinuousDrawingCreatorState extends CreatorState<LineDrawing> {
         /// Draw other points and the whole continuous drawing.
 
         isDrawingFinished = true;
-        final int _currentTap = tapCount - 1;
-        final int _previousTap = tapCount - 2;
 
         edgePoints.add(EdgePoint(
           epoch: epochFromX!(position!.dx),
@@ -80,10 +77,7 @@ class _ContinuousDrawingCreatorState extends CreatorState<LineDrawing> {
         ));
 
         /// Checks if the initial point and the 2nd points are the same.
-        if (Offset(edgePoints[1].epoch.toDouble(),
-                edgePoints[1].quote.toDouble()) ==
-            Offset(edgePoints.first.epoch.toDouble(),
-                edgePoints.first.quote.toDouble())) {
+        if (edgePoints[1] == edgePoints.first) {
           /// If the initial point and the 2nd point are the same,
           /// remove the drawing and clean the drawing tool selection.
           _widget.removeDrawing(drawingId);
@@ -93,30 +87,30 @@ class _ContinuousDrawingCreatorState extends CreatorState<LineDrawing> {
           /// If the initial point and the final point are not the same,
           /// draw the final point and the whole drawing.
           if (tapCount > 2) {
-            drawingId = 'continuous_${edgePoints[_currentTap].epoch}';
-            drawingParts = <LineDrawing>[];
+            drawingParts = <ContinuousLineDrawing>[];
 
-            drawingParts.add(LineDrawing(
+            drawingParts.add(ContinuousLineDrawing(
               drawingPart: DrawingParts.marker,
-              startEdgePoint: edgePoints[_previousTap],
+              startEdgePoint: edgePoints[previousTap],
             ));
           }
-          drawingParts.addAll(<LineDrawing>[
-            LineDrawing(
+          drawingParts.addAll(<ContinuousLineDrawing>[
+            ContinuousLineDrawing(
               drawingPart: DrawingParts.marker,
-              endEdgePoint: edgePoints[_currentTap],
+              endEdgePoint: edgePoints[currentTap],
             ),
-            LineDrawing(
+            ContinuousLineDrawing(
               drawingPart: DrawingParts.line,
-              startEdgePoint: edgePoints[_previousTap],
-              endEdgePoint: edgePoints[_currentTap],
+              startEdgePoint: edgePoints[previousTap],
+              endEdgePoint: edgePoints[currentTap],
             )
           ]);
         }
       }
 
       widget.onAddDrawing(
-        <String, List<LineDrawing>>{drawingId: drawingParts},
+        drawingId,
+        drawingParts,
         isDrawingFinished: isDrawingFinished,
         isInfiniteDrawing: true,
       );
