@@ -1,5 +1,7 @@
+import 'package:equatable/equatable.dart';
+
 /// A class that holds epoch and yCoord of the edge points.
-class EdgePoint {
+class EdgePoint with EquatableMixin {
   /// Initializes
   const EdgePoint({
     this.epoch = 0,
@@ -11,4 +13,7 @@ class EdgePoint {
 
   /// Quote.
   final double quote;
+
+  @override
+  List<Object?> get props => <Object>[epoch, quote];
 }
