@@ -1,4 +1,4 @@
-import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/creator.dart';
+import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/drawing_creator.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/data_model/drawing_parts.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/data_model/edge_point.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/line/line_drawing.dart';
@@ -7,15 +7,10 @@ import 'package:flutter/material.dart';
 /// Creates a Line drawing piece by piece collected on every gesture
 /// exists in a widget tree starting from selecting a line drawing tool and
 /// until drawing is finished
-class LineDrawingCreator extends Creator<LineDrawing> {
+class LineDrawingCreator extends DrawingCreator<LineDrawing> {
   /// Initializes the line drawing creator.
   const LineDrawingCreator({
-    required void Function(
-      Map<String, List<LineDrawing>>, {
-      bool isDrawingFinished,
-      bool isInfiniteDrawing,
-    })
-        onAddDrawing,
+    required OnAddDrawing<LineDrawing> onAddDrawing,
     required double Function(double) quoteFromCanvasY,
     required this.clearDrawingToolSelection,
     required this.removeDrawing,
@@ -33,15 +28,17 @@ class LineDrawingCreator extends Creator<LineDrawing> {
   final void Function(String drawingId) removeDrawing;
 
   @override
-  CreatorState<LineDrawing> createState() => _LineDrawingCreatorState();
+  DrawingCreatorState<LineDrawing> createState() => _LineDrawingCreatorState();
 }
 
-class _LineDrawingCreatorState extends CreatorState<LineDrawing> {
+class _LineDrawingCreatorState extends DrawingCreatorState<LineDrawing> {
   /// If drawing has been started.
   bool _isPenDown = false;
 
   @override
   void onTap(TapUpDetails details) {
+    super.onTap(details);
+
     final LineDrawingCreator _widget = widget as LineDrawingCreator;
 
     if (isDrawingFinished) {
@@ -58,7 +55,6 @@ class _LineDrawingCreatorState extends CreatorState<LineDrawing> {
           quote: widget.quoteFromCanvasY(position!.dy),
         ));
         _isPenDown = true;
-        drawingId = 'line_${edgePoints.first.epoch}';
 
         drawingParts.add(LineDrawing(
           drawingPart: DrawingParts.marker,
@@ -68,8 +64,8 @@ class _LineDrawingCreatorState extends CreatorState<LineDrawing> {
         /// Draw final point and the whole line.
         _isPenDown = false;
         isDrawingFinished = true;
-        final int _currentTap = tapCount - 1;
-        final int _previousTap = tapCount - 2;
+        final int currentTap = tapCount - 1;
+        final int previousTap = tapCount - 2;
 
         edgePoints.add(EdgePoint(
           epoch: epochFromX!(position!.dx),
@@ -77,10 +73,7 @@ class _LineDrawingCreatorState extends CreatorState<LineDrawing> {
         ));
 
         /// Checks if the initial point and the final point are the same.
-        if (Offset(edgePoints[1].epoch.toDouble(),
-                edgePoints[1].quote.toDouble()) ==
-            Offset(edgePoints.first.epoch.toDouble(),
-                edgePoints.first.quote.toDouble())) {
+        if (edgePoints[1] == edgePoints.first) {
           /// If the initial point and the 2nd point are the same,
           /// remove the drawing and clean the drawing tool selection.
           _widget.removeDrawing(drawingId);
@@ -92,12 +85,12 @@ class _LineDrawingCreatorState extends CreatorState<LineDrawing> {
           drawingParts.addAll(<LineDrawing>[
             LineDrawing(
               drawingPart: DrawingParts.marker,
-              endEdgePoint: edgePoints[_currentTap],
+              endEdgePoint: edgePoints[currentTap],
             ),
             LineDrawing(
               drawingPart: DrawingParts.line,
-              startEdgePoint: edgePoints[_previousTap],
-              endEdgePoint: edgePoints[_currentTap],
+              startEdgePoint: edgePoints[previousTap],
+              endEdgePoint: edgePoints[currentTap],
               exceedStart: true,
               exceedEnd: true,
             )
@@ -105,7 +98,8 @@ class _LineDrawingCreatorState extends CreatorState<LineDrawing> {
         }
       }
       widget.onAddDrawing(
-        <String, List<LineDrawing>>{drawingId: drawingParts},
+        drawingId,
+        drawingParts,
         isDrawingFinished: isDrawingFinished,
       );
     });
