@@ -1,4 +1,4 @@
-import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/drawing_creator.dart';
+import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/drawing_tool_widget.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/drawing_data.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/drawing_painter.dart';
 import 'package:deriv_chart/src/deriv_chart/drawing_tool_chart/drawing_tools.dart';
@@ -52,9 +52,10 @@ class DrawingToolChart extends StatelessWidget {
                       quoteFromCanvasY: chartQuoteFromCanvasY,
                       onMoveDrawing: drawingTools.onMoveDrawing,
                       setIsDrawingSelected: _setIsDrawingSelected,
+                      selectedDrawingTool: drawingTools.selectedDrawingTool,
                     )),
             if (drawingTools.selectedDrawingTool != null)
-              DrawingCreator(
+              DrawingToolWidget(
                 onAddDrawing: drawingTools.onAddDrawing,
                 selectedDrawingTool: drawingTools.selectedDrawingTool!,
                 quoteFromCanvasY: chartQuoteFromCanvasY,
